+++ .eslintrc
@@ -24,7 +24,8 @@
   "plugins": [
     "react",
     "import",
-    "jsx-a11y"
+    "jsx-a11y",
+    "flowtype"
   ],
   "settings": {
     "import/resolver": {
