+++ package.json
@@ -160,6 +160,7 @@
     "envalid": "^2.4.1",
     "errorhandler": "^1.4.3",
     "eslint-import-resolver-webpack": "^0.8.0",
+    "eslint-plugin-flowtype": "^2.30.0",
     "eslint-plugin-import": "^2.2.0",
     "express": "^4.14.0",
     "express-promise": "^0.4.0",
